Read sentinel raw dn band data as float64
As all other operations work with float64,
the dn values are also read as float64 (they are actually uint16, but just to make sure all the data has a similar dtype). This should not be a big deal as even 32 bits are enough for sentinel images as far as I managed to research.

New version 7.0.3